src/sidecars/lang-extract/pyproject.toml
[project]
name = "lang-extract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "mcp[cli]>=1.25.0",
    "ollama>=0.4.5",
]
